CUDDFacade: fix bug in inproperly handling variable removal
There was a bug in CUDDFacade::RemoveVariables() method in handling reference
counters of nodes (probably because it internally uses the Apply operation).
It appeared when the removing predicate was removing more than 1 variable. It
is supposed to be fixed now, but because I don't know well the internals of
CUDD, I can't be sure.