OrderedVector: reimplement insert operation for better performance
Original push_back() method that pushed an element at the end of the vector
and then sorted the vector was substituted by insert() method with better
performance. The implementation now attempts to find the element in the
vector using binary search and in case it doesn't find it, the elements are
shifted and the new element is inserted.