Change of value stored in MTBDDs
The values that are stored in leaves of MTBDDs of transition functions are now
not sets of states (for Bottom-Up tree automaton), but sets of special
structures, that behave like

template <typename T>
struct ElemVector
{
  bool isElem;
  union
  {
    T elem;
    SFTA::Vector<T> elemVector;
  }
};

(but are not unions, since SFTA::Vector needs a constructor and therefore
cannot be a part of a union). I hope these all changes work :-).