BUTreeAutomatonCover: add sensible internal states to states translation
The translation of internal states to states is now much more sensible. The
constructor sets the areStatesFromOutside_ variable to true if the automaton
is constructed from outside using some builder... if it is constructed from
the internal automaton representation, the variable is set to false. Depending
on this variable, when the translation is done, it is either done by reverse
lookup in the "state -> internal state" translation table (for the value
"true"), or created automatically (for the value "false")