CompactVariableAssignment: remove problem with allocation of zero-size asgn
This patch removes the bug which occured when a CompactVariableAssignment with
zero variables (i.e.. denoting the universe) was to be allocated. This was
caused by bad computing of number of characters necessary to store such an
assignment, which underflowed to 2^64 which was a requirement that could not
be satisfied by my machine... :-).